requirements.txt
bcrypt==3.2.2
cffi==1.15.0
cryptography==37.0.2
dnspython==2.2.1
ecdsa==0.17.0
email-validator==1.2.1
greenlet==1.1.2
grpcio==1.46.3
grpcio-tools==1.46.3
idna==3.3
passlib==1.7.4
protobuf==3.20.1
psycopg2-binary==2.9.3
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.1
python-dotenv==0.20.0
python-jose==3.3.0
rsa==4.8
six==1.16.0
SQLAlchemy==1.4.37
typing_extensions==4.2.0
